moved some business logic from RawController class  to RawService class and fix some bug in RaceService methods

+++ src/main/java/org/manca/jakarta/project/service/RaceService.java
@@ -16,6 +16,8 @@ public class RaceService {
     @Inject
     private CategoryService categoryService;
     @Inject
+    private AthleteService athleteService;
+    @Inject
     private RaceDao rd;
     @Inject
     RawService rawService;
@@ -34,8 +36,7 @@ public class RaceService {
     }
 
     public boolean addAthlete(Long raceId, Long athleteId, Long categoryID, String raceNumber) {
-        if (raceId!=null && athleteId!=null && categoryID !=null && raceNumber != null
-            && raceId>=1 && athleteId>=1 &&categoryID>=1) {
+        if (this.entitiesExist(raceId, athleteId, categoryID) && raceNumber != null && !raceNumber.equals("")) {
             boolean checkAthleteAdded = rd.addAthlete(raceId, athleteId);
             rawService.setFileName(this.makeName(raceId));
             if(checkAthleteAdded) { //only if persistence operation is successful the RawAthlete will be created and added to the StartList instance via serialization and deserialization operations
@@ -49,7 +50,7 @@ public class RaceService {
     }
 
     public boolean addCategory(Long raceId, Long categoryId) {
-        if (raceId!=null && categoryId!=null && raceId>=1 && categoryId>=1) {
+        if (this.entitiesExist(raceId, categoryId)) {
             boolean checkCategoryAdded =  rd.addCategory(raceId, categoryId);
             rawService.setFileName(this.makeName(raceId));
             if(checkCategoryAdded) { //only if persistence operation is successful the RawCategory will be created and added to the StartList instance via serialization and deserialization operations
@@ -68,11 +69,10 @@ public class RaceService {
 
     public boolean removeCategoryFromRace(Long raceId, Long categoryId) {
         boolean removed = false;
-        if(raceId != null && categoryId != null &&raceId >= 1 && categoryId >= 1) {
+        if(raceId != null && categoryId != null && raceId >= 1 && categoryId >= 1) {
             removed = rd.findById(raceId) != null;
             if(removed) {
-                rawService.setFileName(this.makeName(raceId));
-                removed = rawService.findRawAthleteByCategory(categoryId).size() == 0;
+                removed = rawService.findRawAthleteByCategory(raceId, categoryId).size() == 0;
             }
             if(removed) {
                 removed = rawService.removeRawCategory(categoryId);
@@ -90,13 +90,20 @@ public class RaceService {
      * @return boolean <strong>true</strong> if operation was successful otherwise <strong>false</strong>
      */
     public boolean removeAthleteFromRace(Long raceId, Long athleteId) {
-        if(raceId >= 1 && athleteId >= 1) {
-            rawService.setFileName(this.makeName(raceId));
-            if(rd.removeAthlete(raceId, athleteId))
-                return rawService.removeRawAthlete(athleteId);
+        boolean removed = false;
+
+        if(raceId != null && athleteId != null && raceId >= 1 && athleteId >= 1) {
+            removed = rd.findById(raceId) != null;
+            if (removed) {
+                rawService.setFileName(this.makeName(raceId));
+                removed = rd.removeAthlete(raceId, athleteId);
+                if(removed) {
+                    removed = rawService.removeRawAthlete(athleteId);
+                }
+            }
         }
 
-        return false;
+        return removed;
     }
 
     public List<Category> getAllRaceCategories(Long raceId) {
@@ -138,6 +145,51 @@ public class RaceService {
                 .concat("_")
                 .concat(Integer.toString(rd.findById(raceID).getRaceDateTime().getYear()))
                 .concat(".srd");
+    }
+
+    /**
+     * Validate the ids in three steps, first checks that each of them is not null, then checks that each of them
+     * is greater than 0 and finally verify if a Race, an Athlete and a Category with id that is equal to raceId,
+     * athleteId and categoryId exist in the database respectively
+     * @param raceId the unique Race id
+     * @param athleteId the unique Athlete id
+     * @param categoryId the unique Category id
+     * @return true if the ids constraints are respected otherwise false.
+     * @apiNote 2 overload.
+     */
+    public boolean entitiesExist(Long raceId, Long athleteId, Long categoryId) {
+        return  (raceId != null && athleteId != null && categoryId != null) &&
+                (raceId > 0 && athleteId > 0 && categoryId > 0) &&
+                (this.findRaceById(raceId) != null) &&
+                (athleteService.findAthlete(athleteId)!=null) &&
+                (categoryService.findCategoryById(categoryId)!=null);
+    }
+
+    /**
+     * Validate the ids in three steps, first checks that each of them is not null, then checks that each of them
+     * is greater than 0 and finally verify if both Race and Category with id that is equal to raceId and categoryId
+     * exist in the database respectively.
+     * @param raceId the unique Race id
+     * @param categoryId the unique Category id
+     * @return true if the ids constraints are respected otherwise false.
+     * @apiNote 2 overload.
+     */
+    public boolean entitiesExist(Long raceId, Long categoryId) {
+        return  (raceId != null && categoryId != null) &&
+                (raceId > 0 && categoryId > 0) &&
+                (this.findRaceById(raceId) != null) &&
+                (categoryService.findCategoryById(categoryId)!=null);
+    }
+
+    /**
+     * Validate the id in three steps, first checks that it is not null, then checks that it is greater than 0
+     * and finally verify if Race with id that is equal to raceId, exist in the database.
+     * @param raceId the unique Race id
+     * @return true if the id constraints are respected otherwise false.
+     * @apiNote 2 overload.
+     */
+    public boolean entitiesExist(Long raceId) {
+        return  (raceId != null) && (raceId > 0) && (this.findRaceById(raceId) != null);
     }
 
 }
